+++ docs/src/Box.doc.js
@@ -303,7 +303,7 @@ If you need to use these features for animation purposes, use a \`<div>\` instea
     color="maroon"
     height={50}
   >
-    <Text color="white" weight="bold">Clicking here will do nothing</Text>
+    <Text color="white" weight="bold">Adding onClick here will do nothing</Text>
   </Box>
 </Box>
 `}
@@ -326,7 +326,7 @@ If you need to use these features for animation purposes, use a \`<div>\` instea
         marginTop={3}
       >
         <Text color="white" weight="bold">
-          These all collapsed
+          These margins all collapsed
         </Text>
       </Box>
     </Box>
@@ -796,23 +796,23 @@ card(
     description={`
       [Flex](/Flex)
 
-      - Use Flex for flexbox layouts, especially when even spacing between elements is desired, by using the \`gap\` property.
+      Use Flex for flexbox layouts, especially when even spacing between elements is desired, by using the \`gap\` property.
 
       [Container](/Container)
 
-      - Use Container to responsively layout content with a max-width on large screens.
+      Use Container to responsively layout content with a max-width on large screens.
 
       [ScrollableContainer](/ScrollableContainer)
 
-      - For proper positioning when using anchor components (Flyout, Tooltip, etc.) that can scroll within the viewport, use a ScrollableContainer.
+      For proper positioning when using anchor components (Flyout, Tooltip, etc.) that can scroll within the viewport, use a ScrollableContainer.
 
       [TapArea](/TapArea)
 
-      - If a tap target is needed in order to click on a portion of the page, use TapArea, since \`onClick\` is not supported on Box.
+      If a tap target is needed in order to click on a portion of the page, use TapArea, since \`onClick\` is not supported on Box.
 
       [Sticky](/Sticky)
 
-      - Use Sticky if a portion of the page should stick to either the top or bottom when scrolling.
+      Use Sticky if a portion of the page should stick to either the top or bottom when scrolling.
 
     `}
   />,