+++ docs/src/Button.doc.js
@@ -226,6 +226,16 @@ card(
         ],
         href: 'type-roles',
       },
+      {
+        name: 'onNavigationOptions',
+        type: '({ [string]: Node | ({| +event: SyntheticEvent<> |}) => void }) => void',
+        description: [
+          'onNavigationOptions works in conjunction with a Provider. Pass custom props to onNavigation. See Provider for examples.',
+          `onNavigation's type is flexible. Each key's value is a React.Node or an event handler function.`,
+          'Optional with role=link.',
+        ],
+        href: 'OnNavigationContext',
+      },
     ]}
   />,
 );