+++ packages/gestalt/src/MenuOption.js
@@ -12,6 +12,10 @@ import getRoundingClassName from './getRoundingClassName.js';
 import Icon from './Icon.js';
 import focusStyles from './Focus.css';
 import useFocusVisible from './useFocusVisible.js';
+import {
+  type OnNavigationOptionsType,
+  OnNavigationOptionsPropType,
+} from './contexts/OnNavigation.js';
 
 export type OptionObject = {|
   label: string,
@@ -38,6 +42,7 @@ type Props = {|
   shouldTruncate?: boolean,
   textWeight?: FontWeight,
   href?: string,
+  onNavigationOptions?: OnNavigationOptionsType,
 |};
 
 export default function MenuOption({
@@ -56,6 +61,7 @@ export default function MenuOption({
   shouldTruncate = false,
   textWeight = 'normal',
   href,
+  onNavigationOptions,
 }: Props): Node {
   const matches = (Array.isArray(selected) ? selected : []).filter(
     ({ value }) => value === option.value,
@@ -158,7 +164,12 @@ export default function MenuOption({
     >
       <Box padding={2} color={optionStateColor} rounding={2} display="flex" direction="column">
         {href ? (
-          <Link hoverStyle="none" href={href} target="blank">
+          <Link
+            hoverStyle="none"
+            href={href}
+            onNavigationOptions={onNavigationOptions ?? {}}
+            target="blank"
+          >
             {menuOptionContents}
           </Link>
         ) : (
@@ -199,4 +210,6 @@ MenuOption.propTypes = {
   hoveredItem: PropTypes.number,
   setHoveredItem: PropTypes.func,
   setOptionRef: PropTypes.func,
+  href: PropTypes.string,
+  onNavigationOptions: OnNavigationOptionsPropType,
 };