+++ packages/gestalt/src/commonTypes.js
@@ -1,6 +1,10 @@
 // @flow strict
 import PropTypes from 'prop-types';
 import { type AbstractEventHandler } from './AbstractEventHandler.js';
+import {
+  type OnNavigationOptionsType,
+  OnNavigationOptionsPropType,
+} from './contexts/OnNavigation.js';
 
 export type ActionDataType = {|
   accessibilityLabel?: string,
@@ -12,6 +16,7 @@ export type ActionDataType = {|
     | SyntheticKeyboardEvent<HTMLAnchorElement>
     | SyntheticKeyboardEvent<HTMLButtonElement>,
   >,
+  onNavigationOptions?: OnNavigationOptionsType,
 |};
 
 export type DismissButtonType = {|
@@ -26,6 +31,7 @@ export const ActionDataPropType: React$PropType$Primitive<ActionDataType> = Prop
   // $FlowFixMe[incompatible-type]
   onClick: PropTypes.func,
   accessibilityLabel: PropTypes.string,
+  onNavigationOptions: OnNavigationOptionsPropType,
 });
 // $FlowFixMe[incompatible-exact]
 // $FlowFixMe[incompatible-type]